fix: add boar content writer object

+++ backend/src/main/java/com/ssafy/doyouwannabuildasnowball/service/BoardService.java
@@ -2,12 +2,14 @@ package com.ssafy.doyouwannabuildasnowball.service;
 
 import com.ssafy.doyouwannabuildasnowball.common.exception.CustomException;
 import com.ssafy.doyouwannabuildasnowball.domain.Board;
+import com.ssafy.doyouwannabuildasnowball.domain.Member;
 import com.ssafy.doyouwannabuildasnowball.domain.Snowglobe;
 import com.ssafy.doyouwannabuildasnowball.dto.board.BoardDto;
 import com.ssafy.doyouwannabuildasnowball.dto.board.request.WriteBoardRequest;
 import com.ssafy.doyouwannabuildasnowball.dto.board.response.BoardAllResponse;
 import com.ssafy.doyouwannabuildasnowball.dto.board.response.BoardResponse;
 import com.ssafy.doyouwannabuildasnowball.repository.jpa.BoardRepository;
+import com.ssafy.doyouwannabuildasnowball.repository.jpa.MemberRepository;
 import com.ssafy.doyouwannabuildasnowball.repository.jpa.SnowglobeRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -16,6 +18,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static com.ssafy.doyouwannabuildasnowball.common.exception.ErrorCode.*;
 
@@ -28,6 +31,7 @@ public class BoardService {
     private final BoardRepository boardRepository;
 
     private final SnowglobeRepository snowglobeRepository;
+    private final MemberRepository memberRepository;
 
     public BoardAllResponse findAllContentsBySnowglobe(Long snowglobeId) {
 
@@ -41,6 +45,7 @@ public class BoardService {
                         .modifiedTime(board.getModifiedTime())
                         .boardId(board.getBoardId())
                         .snowglobeId(board.getSnowglobe().getSnowglobeId())
+                        .writerId(board.getWriter().getMemberId())
                         .content(board.getContent())
                         .imageUrl(board.getPicture()).build()
         ));
@@ -63,6 +68,7 @@ public class BoardService {
                         .content(writeBoardRequest.getContent())
                         .picture(imageURL)
                         .snowglobe(snowglobe)
+                        .writer(memberRepository.findById(writeBoardRequest.getWriterId()).orElse(null))
                         .build());
     }
 
@@ -71,14 +77,21 @@ public class BoardService {
 
         Board board = boardRepository.findById(boardDto.getBoardId())
                 .orElseThrow(() -> new CustomException(BOARD_NOT_FOUND));
+        if(board.getWriter().getMemberId() != boardDto.getWriterId()) {
+            throw new CustomException(UNMATCHED_MEMBER);
+        }
         String imageURL = boardDto.getPicture();
         board.contentUpdate(boardDto.getContent(), imageURL);
         boardRepository.updateBoardContent(board.getContent(), board.getPicture(), board.getBoardId());
 
     }
 
-    public void removeContent(Long boardId) {
+    public void removeContent(Long boardId, Long memberId) {
         log.info("board id : " + boardId);
+        Board board = boardRepository.findById(boardId).orElseThrow(() -> new CustomException(BOARD_NOT_FOUND));
+        Long boardOwnerId = board.getSnowglobe().getReceiver().getMemberId();
+        if(!Objects.equals(memberId, boardOwnerId) || !Objects.equals(memberId, board.getWriter().getMemberId()))
+            throw new CustomException(MEMBER_NOT_FOUND);
         boardRepository.deleteById(boardId);
     }
 }
